# Conflicts:
#	.gitignore
#	src/main/java/Map/Building.java
#	src/main/java/Map/Floor.java
#	src/main/java/Map/Location.java
#	src/main/java/Map/Node.java
#	src/test/java/MapTest/BuildingTest.java

+++ src/main/java/Map/Location.java
@@ -1,14 +1,12 @@
 package Map;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 /**
  * Class that contains the x and y coordinates of a arbitrary point.
  */
-
 public class Location {
-    private int x;
-    private int y;
+
+    private double x;
+    private double y;
 
     /**
      * Default constructor for Location Class
@@ -16,7 +14,7 @@ public class Location {
      * @param x x-coordinate
      * @param y y-coordinate
      */
-    public Location(int x, int y) {
+    public Location(double x, double y) {
 
         this.x = x;
         this.y = y;
@@ -28,7 +26,7 @@ public class Location {
      *
      * @return The location's x-coordinate
      */
-    public int getX() {
+    public double getX() {
 
         return x;
     }
@@ -38,7 +36,7 @@ public class Location {
      *
      * @return The location's y-coordinate
      */
-    public int getY() {
+    public double getY() {
 
         return y;
     }
@@ -48,7 +46,7 @@ public class Location {
      *
      * @param x The new x-coordinate
      */
-    public void setX(int x) {
+    public void setX(double x) {
 
         this.x = x;
 
@@ -59,7 +57,7 @@ public class Location {
      *
      * @param y The new y-coordinate
      */
-    public void setY(int y) {
+    public void setY(double y) {
 
         this.y = y;
